Move *Report class to the report package
