+++ src/main/java/io/vertx/ext/unit/impl/TestSuiteReportImpl.java
@@ -3,9 +3,9 @@ package io.vertx.ext.unit.impl;
 import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.streams.ReadStream;
-import io.vertx.ext.unit.TestSuiteReport;
+import io.vertx.ext.unit.report.TestSuiteReport;
 import io.vertx.ext.unit.Test;
-import io.vertx.ext.unit.TestCaseReport;
+import io.vertx.ext.unit.report.TestCaseReport;
 
 /**
 * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
