+++ src/main/java/io/vertx/ext/unit/report/TestCaseReport.java
@@ -1,13 +1,10 @@
-package io.vertx.ext.unit;
+package io.vertx.ext.unit.report;
 
 import io.vertx.codegen.annotations.CacheReturn;
 import io.vertx.codegen.annotations.Fluent;
-import io.vertx.codegen.annotations.GenIgnore;
 import io.vertx.codegen.annotations.VertxGen;
 import io.vertx.core.Handler;
-import io.vertx.core.Vertx;
-
-import java.util.concurrent.TimeUnit;
+import io.vertx.ext.unit.TestResult;
 
 /**
  * @author <a href="mailto:julien@julienviet.com">Julien Viet</a>
