+++ src/main/resources/vertx-unit-js/test_case_report.js
@@ -20,7 +20,7 @@ var TestResult = require('vertx-unit-js/test_result');
 
 var io = Packages.io;
 var JsonObject = io.vertx.core.json.JsonObject;
-var JTestCaseReport = io.vertx.ext.unit.TestCaseReport;
+var JTestCaseReport = io.vertx.ext.unit.report.TestCaseReport;
 
 /**
